examples/02_golf_coach_example/pyproject.toml
[project]
name = "stream-agents-golf"
version = "0.0.0"
requires-python = ">=3.13"

# put only what this example needs
dependencies = [
  "python-dotenv>=1.0",
  "stream-agents",
  "ultralytics>=8.3.184",
  "av>=14.4.0",
]

[tool.uv.sources]
"stream-agents" = {path = "../../agents-core", editable=true}
